+++ client/mgmt.h
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+/*
+ *
+ *  BlueZ - Bluetooth protocol stack for Linux
+ *
+ *  Copyright (C) 2023 Intel Corporation. All rights reserved.
+ *
+ *
+ */
+
+bool mgmt_add_submenu(void);
+void mgmt_remove_submenu(void);
+void mgmt_set_index(const char *arg);
